--- src/pages/Login/Login.scss
@@ -1,55 +1,0 @@
-.login_container {
-    height: 80vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin: auto 25%;
-}
-
-.login_input {
-    margin-bottom: 2%;
-    width: 80%;
-}
-
-.chakra-input__group.css-1x2eqe {
-    display: flex;
-    align-content: space-around;
-}
-
-.input_container {
-    border: 4px solid #EDF2F4;
-    padding: 5% 5% 5% 5%;
-    border-radius: 3%;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    width: 50%;
-    margin: auto;
-}
-
-.mobile_login_container {
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin: -30% 10% 0%;
-}
-
-.mobile_login_input {
-    margin-bottom: 3%;
-    margin-left: 5%;
-    padding: 2% 3%;
-}
-
-.mobile_input_container {
-    border: 4px solid #EDF2F4;
-    padding: 2%;
-    border-radius: 5%;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-}
-
-.mobile_login_button {
-    padding-left: 8%;
-}
