+++ todo.md
@@ -1,1 +1,2 @@
 -> voir pour inclure l'appel à delete refresh_token dans la fonction logout
+->Modifier le login pour retirer l'envoie du role
